+++ kite/src/main/kotlin/com/tang/kite/result/time/TimeResultHandler.kt
@@ -13,9 +13,9 @@ class TimeResultHandler : ResultHandler {
 
     override fun <T> setValue(field: Field, instance: T, value: Any) {
         when (value) {
-            is Time -> field.set(instance, value.toLocalTime())
-            is LocalTime -> field.set(instance, value)
-            is LocalDateTime -> field.set(instance, value.toLocalTime())
+            is Time -> field.set(instance, value)
+            is LocalTime -> field.set(instance, Time.valueOf(value))
+            is LocalDateTime -> field.set(instance, Time.valueOf(value.toLocalTime()))
             else -> throw IllegalArgumentException("Unsupported type: ${value::class.java.name} for field: ${field.name}")
         }
     }